[Add] espnet2 core

+++ espnet2/layers/stft.py
@@ -101,9 +101,8 @@ class Stft(torch.nn.Module, InversibleInterface):
                 onesided=self.onesided,
             )
             stft_kwargs["return_complex"] = True
-            # NOTE(Jinchuan) CuFFT is not compatible with bfloat16
-            output = torch.stft(input.float(), **stft_kwargs)
-            output = torch.view_as_real(output).type(input.dtype)
+            output = torch.stft(input, **stft_kwargs)
+            output = torch.view_as_real(output)
         else:
             if self.training:
                 raise NotImplementedError(
